Refactor ArrowEffects::GetTime to return microseconds
Changed ArrowEffects::GetTime to return time in microseconds as uint64_t instead of float seconds. Updated all usages to convert microseconds to seconds where needed, improving precision and consistency in time calculations for arrow effects.

+++ src/ArrowEffects.h
@@ -11,7 +11,7 @@ class ArrowEffects
 {
 public:
 	static void Init(PlayerNumber pn);
-	static float GetTime();
+	static uint64_t GetTime();
 	static void Update();
 	// SetCurrentOptions and the hidden static variable it set exists so that
 	// ArrowEffects doesn't have to reach through the PlayerState to check
